+++ src/Cwaff-Misc/CwaffRunData.cs
@@ -6,6 +6,7 @@ public class CwaffRunData : FakeItem
     public string btcktfEnemyGuid = string.Empty;
     public bool shouldReturnToPreviousFloor = false;
     public string nameOfPreviousFloor = null;
+    public List<int>[] glassGunIds = [new(), new()];
 
     private bool _deserialized = false;
 
@@ -48,17 +49,41 @@ public class CwaffRunData : FakeItem
     public override void MidGameSerialize(List<object> data)
     {
         base.MidGameSerialize(data);
+
         data.Add(btcktfEnemyGuid);
+
+        data.Add(glassGunIds[0].Count);
+        for (int i = 0; i < glassGunIds[0].Count; ++i)
+            data.Add(glassGunIds[0][i]);
+
+        data.Add(glassGunIds[1].Count);
+        for (int i = 0; i < glassGunIds[1].Count; ++i)
+            data.Add(glassGunIds[1][i]);
     }
 
     public override void MidGameDeserialize(List<object> data)
     {
         Lazy.DebugLog($"deserializing extra mid run data");
+        _Instance = this;
         base.MidGameDeserialize(data);
         int i = 0;
 
         btcktfEnemyGuid = (string)data[i++];
         Lazy.DebugLog($"  memorialized enemy is {btcktfEnemyGuid}");
+
+        int p1NumGlassGuns = (int)data[i++];
+        glassGunIds[0] = new List<int>();
+        for (int gunNum = 0; gunNum < p1NumGlassGuns; ++gunNum)
+            glassGunIds[0].Add((int)data[i++]);
+        if (p1NumGlassGuns > 0 && GameManager.Instance.PrimaryPlayer is PlayerController p1)
+            GlassAmmoBox.RestoreMidGameData(p1);
+
+        int p2NumGlassGuns = (int)data[i++];
+        glassGunIds[1] = new List<int>();
+        for (int gunNum = 0; gunNum < p2NumGlassGuns; ++gunNum)
+            glassGunIds[1].Add((int)data[i++]);
+        if (p2NumGlassGuns > 0 && GameManager.Instance.SecondaryPlayer is PlayerController p2)
+            GlassAmmoBox.RestoreMidGameData(p2);
 
         _Instance = this;
         this._deserialized = true;